Added changeable distance functions

+++ src/edu/hendrix/modeselection/cluster/DistanceFunc.java
@@ -1,7 +1,7 @@
 package edu.hendrix.modeselection.cluster;
 
 public interface DistanceFunc<T> {
-	public long distance(T img1, T img2);
+	public double distance(T img1, T img2);
 	
 	default public long square(long value) {
 		return value * value;
